chore[upa]: move test deploy utils to their own file
